display image in services
