+++ app/src/main/java/org/ole/planet/myplanet/utilities/DownloadUtils.kt
@@ -1,6 +1,8 @@
 package org.ole.planet.myplanet.utilities
 
 import org.ole.planet.myplanet.model.RealmMyLibrary
+import java.util.regex.Pattern
+import kotlin.text.isNotEmpty
 
 object DownloadUtils {
     @JvmStatic
@@ -19,5 +21,22 @@ object DownloadUtils {
             urls.add(Utilities.getUrl(dbMyLibrary[selectedItems[i]]))
         }
         return urls
+    }
+
+    fun extractLinks(text: String?): ArrayList<String> {
+        val links = ArrayList<String>()
+        val pattern = Pattern.compile("!\\[.*?]\\((.*?)\\)")
+        val matcher = text?.let { pattern.matcher(it) }
+        if (matcher != null) {
+            while (matcher.find()) {
+                val link = matcher.group(1)
+                if (link != null) {
+                    if (link.isNotEmpty()) {
+                        links.add(link)
+                    }
+                }
+            }
+        }
+        return links
     }
 }
